orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "Startup in a Box — Orchestrator (Google ADK + ag-ui-adk) scaffold"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "pydantic>=2.8",
    "pydantic-settings>=2.5",
    "python-dotenv>=1.0",
    "google-adk>=0.3.0",
    "ag-ui-adk==0.6.0",
    "ag-ui-protocol==0.1.15",
    "litellm>=1.83.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "asgi-lifespan>=2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/orchestrator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
